Split the classes in graphics.cpp into different scripts and added header files to all graphics scripts. Also began setting up graphics in main, but it is not compatible with the other test code yet.
