+++ src/player.cpp
@@ -22,6 +22,6 @@ void player::updateDirection()
 
 void player::onRender()
 {
-    Rectangle rectangle = bounds();
-    DrawRectangle(rectangle.x, rectangle.y, rectangle.width, rectangle.height, GREEN);
+    //Rectangle rectangle = bounds();
+    //DrawRectangle(rectangle.x, rectangle.y, rectangle.width, rectangle.height, GREEN);
 }
