+++ src/textureloader.cpp
@@ -2,15 +2,14 @@
 * Adam Aronow
 */
 
-#pragma once
-
-#include <iostream>
-#include <unordered_map>
-#include "raylib.h"
+#include "textureLoader.h"
 
 extern std::unordered_map<std::string, Texture2D> textureMap;
 
-static void loadAllTextures( )
+/*-----------------------------------------------------------------------------------------
+* @brief: loads every texture and assigns it to the textureMap with a corresponding string
+*/
+void loadAllTextures( )
 {
 	textureMap[ "player" ] = LoadTexture( ( char* ) "../../../assets/graphics/placeholders/Placeholder player down (1).png" );
 	textureMap[ "enemy" ] = LoadTexture( ( char* ) "../../../assets/graphics/placeholders/Placeholder enemy (1).png" );
@@ -19,7 +18,10 @@ static void loadAllTextures( )
 	textureMap[ "floor" ] = LoadTexture( ( char* ) "../../../assets/graphics/placeholders/Placeholder floor tile.png" );
 }
 
-static void unloadAllTextures( )
+/*---------------------------------------------
+* @brief: unloads every texture in textureMap
+*/
+void unloadAllTextures( )
 {
 	for ( auto i = textureMap.begin( ); i != textureMap.end( ); i++ )
 	{
